revert false positive

+++ board/safety/safety_volkswagen_pq.h
@@ -70,7 +70,7 @@ static uint32_t volkswagen_pq_compute_checksum(CANPacket_t *to_push) {
   int addr = GET_ADDR(to_push);
   int len = GET_LEN(to_push);
   uint8_t checksum = 0U;
-  int checksum_byte = (addr == MSG_MOTOR_5) ? (int) 7 : (int) 0;
+  int checksum_byte = (addr == MSG_MOTOR_5) ? 7 : 0;
 
   // Simple XOR over the payload, except for the byte where the checksum lives.
   for (int i = 0; i < len; i++) {
